Improve Graph Layout and Address Data (#2)
This patch runs the force-directed layout of the map labels asynchronously, meaning that the UI doesn't hang while we're calculating the label positions. I may revisit this in the future, but I think it is generally an improvement. I ported the leaflet-tooltip-layout package to TypeScript to do this, and then made some changes to it.

I also exported the contents of the sqlite database as a set of JSON files, one per table. The goal is to eventually cut out the sqlite file all together, and cut out the entire api package as well. We can instead use webpack to stuff the JSON files directly into our bundle. This will cut down on the development overhead of generating static API files, and in production this means that all the data can be served as a single bundle, instead of making numerous requests to fetch data. There's not really that much data, so it shouldn't be a huge deal.

I also refactored the way that addresses are displayed to show where the information came from. At the moment, this could be (1) RFC 518, (2) the 1982 ARPANET Directory, or (3) manual entry.

I added a new table called `gateway_metadata` for associating long names and descriptions with gateway acronyms.